src/pyproject.toml
[project]
name = "graphwiki"
version = "0.1.0"
description = "A modern wiki platform inspired by MoinMoin, Graphingwiki, and Obsidian"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graphwiki"]
